fixing problems, gui improvement, help texts added, copyright notices

+++ resources/views/admin/user.blade.php
@@ -1,3 +1,6 @@
+
+<!-- # (C) 2016 Matthias Kuhs, Ireland -->
+
 @extends('layouts.main')
 
 @section('title', "Create or Update a User")
@@ -21,32 +24,42 @@
 
         <p>{!! Form::label('first_name', 'First Name'); !!}<br>
            {!! Form::text('first_name'); !!}</p>
+
         <p>{!! Form::label('last_name', 'Last Name'); !!}<br>
            {!! Form::text('last_name'); !!}</p>
+
         <p>{!! Form::label('email', 'Email Address'); !!}<br>
-           {!! Form::text('email'); !!}</p>
-        <strong>Select Roles:</strong><br />
-        @foreach ($roles as $role)
-            <input name="{{ $role->name }}" type="checkbox"
-                 {{ isset($user) && $user->hasRole($role->name) ? 'checked="checked"' : '' }}>
-            <label  for="{{ $role->name }}">{{ $role->name }}</label><br />
-        @endforeach
+        @if (Auth::user()->isAdmin())
+               {!! Form::text('email'); !!}</p>
+
+            <strong>Select Roles:</strong><br />
+            @foreach ($roles as $role)
+                <input name="{{ $role->name }}" type="checkbox"
+                     {{ isset($user) && $user->hasRole($role->name) ? 'checked="checked"' : '' }}>
+                <label  for="{{ $role->name }}">{{ $role->name }}</label><br />
+            @endforeach
+        @else
+            {!! Form::hidden('email'); !!}
+            {{ $user->email }}
+        @endif
 
 
     @if (isset($user))
         <p>{!! Form::submit('Update'); !!}</p>
-        <hr>
-        <a class="btn btn-danger btn-sm"  role="button" href="/admin/users/{{ $user->id }}/delete">
-            <i class="fa fa-trash" > </i> &nbsp; Delete
-        </a>
+        @if (Auth::user()->isAdmin())
+            <hr>
+            <a class="btn btn-danger btn-sm"  role="button" href="/admin/users/{{ $user->id }}/delete">
+                <i class="fa fa-trash" > </i> &nbsp; Delete
+            </a>
+        @endif
     @else
-        <p>{!! Form::submit('Submit'); !!}
+        {!! Form::submit('Submit'); !!}
     @endif
 
-    <script type="text/javascript">document.forms.inputForm.first_name.focus()</script>
-
-    <a href="/admin/users">{!! Form::button('Cancel'); !!}</a></p>
+    <a href="#" onclick="history.back()">{!! Form::button('Cancel'); !!}</a>
     {!! Form::close() !!}
+
+    <script type="text/javascript">document.forms.inputForm.first_name.focus()</script>
     
 @stop
